// src/types/api/petType.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * Pet Management API
 * Pet Management API with health logs, prescriptions, item master, and JWT authentication.
 * OpenAPI spec version: 2
 */

export type PetType = (typeof PetType)[keyof typeof PetType];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const PetType = {
  dog: "dog",
  cat: "cat",
  rabbit: "rabbit",
  hamster: "hamster",
  bird: "bird",
  turtle: "turtle",
  fish: "fish",
} as const;